# user-microservice/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: User Microservice
  description: |-
    This is the User Microservice for the EasyConf scenario in CSE2115. It is responsible for managing users, events (including their tracks and attendees), and invitations.
  version: 0.0.0
servers:
  - url: http://localhost:8085
    description: Production server for integration.
...
paths:
...
  /event:
    put:
      tags:
        - Event Management
      summary: Update an existing event.
      description: This operation can only be performed by the general chair of the event.
      operationId: updateEvent
      requestBody:
        description: Event data to be changed.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
        required: true
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid event object supplied.
        '401':
          description: Unauthorized access.
        '404':
          description: Event not found.
      security:
        - api_key: [ ]
    post:
      tags:
        - Event Management
      summary: Create a new event.
      description: Upon creation, the user associated with the provided token will be designated the general chair of the event.
      operationId: addEvent
      requestBody:
        description: Event to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
        required: true
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          description: Unauthorized access.
        '400':
          description: Invalid input.
      security:
        - api_key: [ ]
    get:
      tags:
        - Event Management
      summary: Finds Events
      description: Finds Events
      operationId: findEvent
      parameters:
        - name: start_before
          in: query
          description: Date before which event starts
          required: false
          explode: true
          schema:
            type: string
            format: date
        - name: start_after
          in: query
          description: Date after which event starts
          required: false
          explode: true
          schema:
            type: string
            format: date
        - name: end_before
          in: query
          description: Date before which event ends
          required: false
          explode: true
          schema:
            type: string
            format: date
        - name: end_after
          in: query
          description: Date after which event ends
          required: false
          explode: true
          schema:
            type: string
            format: date
        - name: cancelled
          in: query
          description: Whether the event is cancelled
          required: false
          explode: true
          schema:
            type: boolean
            default: false
        - name: name
          in: query
          description: Event's name.
          required: false
          explode: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Invalid query values.
        '401':
          description: Unauthorized access.
      security:
        - api_key: [ ]
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 4321234
        start_date:
          type: string
          format: date
          example: 2001-07-04 12:08
        end_date:
          type: string
          format: date
          example: 2001-08-05 15:23
        is_cancelled:
          type: boolean
        name:
          type: string
          example: Open conference for paper review
        description:
          nullable: true
          type: string
          example: this is a conference that deals with submitting papers to review.
...
  securitySchemes:
    api_key:
      type: http
      scheme: bearer
      bearerFormat: JWT